Add Mapbox components and configuration files
Added new Mapbox components (MapboxContainer.vue and MapboxGL.vue) along with necessary configuration files (.gitignore, tsconfig.app.json, and pnpm-lock.yaml). This will enable the use of Mapbox for map-related functionalities in the application. Settings related to initial map location, as well as the reset and update functions, are configured.