Change Mapbox location and add new Position type
The Position type was added to the MapboxContainer component, improving the readability and maintainability of the code. The initial location settings were refactored to use the new Position type and the latitude and longitude were changed for the base position. Furthermore, the resetLocation function was updated to reset location to the base position.